requirements.txt
asgiref==3.5.2
Django==3.2.15
psycopg2-binary==2.9.3
sqlparse==0.4.2
django_ckeditor==6.5.1
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6